navigation standard added
