--- navigation/packages/nav_main/launch/task_launch/storing_groceries.launch.py
@@ -1,60 +1,0 @@
-import os
-from ament_index_python.packages import get_package_share_directory
-from launch import LaunchDescription
-from launch_ros.actions import Node
-from launch.actions import IncludeLaunchDescription, DeclareLaunchArgument,OpaqueFunction
-from launch.launch_description_sources import PythonLaunchDescriptionSource
-from launch.substitutions import LaunchConfiguration, PathJoinSubstitution
-from launch_ros.substitutions import FindPackageShare
-
-def launch_setup(context, *args, **kwargs):
-    nav_dir = get_package_share_directory('nav_main')
-    use_sim = LaunchConfiguration('use_sim', default='false')
-    localization = LaunchConfiguration('localization', default='true')
-    rtabmap_viz = LaunchConfiguration('rtabmap_viz', default='false')
-    default_value=os.path.join(nav_dir, 'config', 'new_params.yaml'),
-    params_file = LaunchConfiguration('params_file', default=default_value)
-    
-    
-    nav_basics = IncludeLaunchDescription(
-        PythonLaunchDescriptionSource(
-            PathJoinSubstitution(
-                [
-                    FindPackageShare("nav_main"),
-                    "launch",
-                    "nav_basics.launch.py",
-                ]
-            )),
-        launch_arguments={'use_sim': use_sim,}.items()
-        )
-    nav2_launch = IncludeLaunchDescription(
-        PythonLaunchDescriptionSource(
-            PathJoinSubstitution(
-                [
-                    FindPackageShare("nav_main"),
-                    "launch",
-                    "nav2.launch.py",
-                ]
-            )),
-        launch_arguments={'use_sim_time': use_sim, 'params_file': params_file}.items()
-        )
-    rtabmap = IncludeLaunchDescription(
-        PythonLaunchDescriptionSource(
-            PathJoinSubstitution(
-                [
-                    FindPackageShare("nav_main"),
-                    "launch",
-                    "rtabmap_slam.launch.py",
-                ]
-            )),
-        launch_arguments={'use_sim_time': use_sim, 'localization': localization, 'rtabmap_viz': rtabmap_viz}.items()
-        )
-    
-    return [
-        nav_basics,
-        rtabmap,
-        nav2_launch,
-    ]
-
-def generate_launch_description():
-    return LaunchDescription([OpaqueFunction(function=launch_setup)])
